getProfileInfo api코드 작성 & 타입 이동
