+++ src/entities/user/api/index.ts
@@ -1,0 +1,1 @@
+export { getProfileInfo } from "./getProfileInfo";
