+++ src/entities/user/index.ts
@@ -1,1 +1,2 @@
 export * from "./model";
+export * from "./api";
