+++ src/entities/user/model/type.ts
@@ -2,3 +2,16 @@ export type SignInUserInfoType = {
   userId: string;
   password: string;
 };
+
+export type ProfileType = {
+  id: string;
+  name: string;
+  job: string;
+  location: string;
+  imgUrl: string;
+  portfolio?: string;
+  email?: string;
+  linkedinUrl?: string;
+  gitHubUrl?: string;
+  introduce?: string;
+};
