+++ src/features/Profile/index.ts
@@ -1,2 +1,1 @@
 export { ProfileAside } from "./ui/ProfileAside";
-export * from "./model";
