--- src/features/Profile/model/type.ts
@@ -1,12 +1,0 @@
-export type ProfileType = {
-  id: string;
-  name: string;
-  job: string;
-  location: string;
-  imgUrl: string;
-  portfolio?: string;
-  email?: string;
-  linkedinUrl?: string;
-  gitHubUrl?: string;
-  introduce?: string;
-};
